Better display of nutrimeal and nutricast data
Fixed issues with the display of nutricast and nutrimeal data in the frontend application. The changes ensure that the data is presented more clearly and is easier to read.
Now, the nutritional information is formatted consistently, and the layout is improved for better user experience.
It also fixes one UI bug, where the goals values were not displaying correctly in the nutrimeal view.